requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.2.3
attrs==23.2.0
blinker==1.8.2
blis==0.7.11
boto3==1.34.144
botocore==1.34.144
cachetools==5.4.0
catalogue==2.0.10
cement==3.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cmake==3.27.7
colorama==0.4.6
confection==0.1.5
crispy-bootstrap4==2024.1
cryptography==42.0.8
cymem==2.0.8
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.1.0
Django==5.0.7
django-allauth==0.63.6
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cleanup==8.1.0
django-countries==7.6.1
django-crispy-forms==2.2
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==24.2
django-heroku==0.3.1
django-jalali==7.0.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-pwa==1.1.0
django-ranged-response==0.2.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-simple-captcha==0.6.0
django-storages==1.14.4
django-taggit==5.0.1
django-user-visit==2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
docutils==0.21.2
docx2txt==0.8
et-xmlfile==1.1.0
filelock==3.13.1
flake8==3.8.4
Flask==3.0.3
Flask-MySQLdb==2.0.0
fsspec==2024.6.1
future==1.0.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gTTS==2.5.1
gTTS-token==1.1.4
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.10.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
lazy-object-proxy==1.10.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.6.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
mysqlclient==2.2.4
networkx==3.3
nltk==3.8.1
numpy==1.26.2
oauth2==1.9.0.post1
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.3
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
pathspec==0.12.1
patsy==0.5.6
pdfminer.six==20240706
pgmpy==0.1.25
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.0.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
psycopg2-binary==2.8.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==2.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
setuptools==69.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
social-auth-core==4.5.4
sortedcontainers==2.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
statsmodels==0.14.2
sympy==1.13.1
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
toml==0.10.2
torch==2.4.0
tqdm==4.66.4
typed_ast==1.5.5
typer==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.2.2
user-agents==2.2.0
virtualenv==20.24.7
virtualenvwrapper-win==1.2.7
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.3
whitenoise==4.1.2
wrapt==1.16.0
